subindo projeto

+++ src/main/java/com/unique/events/register/domain/Desktop.java
@@ -10,22 +10,22 @@ public class Desktop {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Integer id_desktop;
+	private long id_desktop;
 
 	public Desktop() {
 		super();
 	}
 
-	public Desktop(Integer id_desktop) {
+	public Desktop(long id_desktop) {
 		super();
 		this.id_desktop = id_desktop;
 	}
 
-	public Integer getId_desktop() {
+	public long getId_desktop() {
 		return id_desktop;
 	}
 
-	public void setId_desktop(Integer id_desktop) {
+	public void setId_desktop(long id_desktop) {
 		this.id_desktop = id_desktop;
 	}
